+++ app/feature/home/constants/banner-keyword.ts
@@ -1,0 +1,24 @@
+import ElectricCordIcon from "@common/assets/icons/electric-cord/ElectricCordIcon";
+import HeartIcon from "@common/assets/icons/heart/HeartIcon";
+import LampIcon from "@common/assets/icons/lamp/LampIcon";
+import MusicIcon from "@common/assets/icons/music/MusicIcon";
+import PetIcon from "@common/assets/icons/pet/PetIcon";
+import WindowIcon from "@common/assets/icons/window/WindowIcon";
+import WineIcon from "@common/assets/icons/wine/Wine";
+
+export const HEADER_KEYWORD = [
+  { icon: HeartIcon, label: "연인과의 데이트" },
+  { icon: MusicIcon, label: "잔잔한 음악" },
+  { icon: LampIcon, label: "따뜻한 조명" },
+  { icon: WineIcon, label: "와인" },
+  { icon: ElectricCordIcon, label: "콘센트 있음" },
+  { icon: PetIcon, label: "애견동반 가능" },
+  { icon: WindowIcon, label: "통창뷰" },
+  { icon: HeartIcon, label: "연인과의 데이트" },
+  { icon: MusicIcon, label: "잔잔한 음악" },
+  { icon: LampIcon, label: "따뜻한 조명" },
+  { icon: WineIcon, label: "와인" },
+  { icon: ElectricCordIcon, label: "콘센트 있음" },
+  { icon: PetIcon, label: "애견동반 가능" },
+  { icon: WindowIcon, label: "통창뷰" },
+];
